+++ Documentation/CHANGELOG.md
@@ -2,21 +2,27 @@ Version control history:
 ====================
 
 * Version 1.0-1 020717
-	* First version
+	* First version 
 
 * Version 1.1-2 220717
 	* CPU Alarm function added
 
-* Version 1.2-3 120817
+* Version 1.2-3 040817
 	* GPIO LED feature added.
 
 * Version 1.3-4 140817
 	* Graph of logfile function added.
 	
-* Version 1.4-5 200817
+* Version 1.4-5 290817
 	* Code rewritten complete in python3 (from a mixture of bash and python)
 	* Graph of realtime CPU data added
 
 * Version 1.5-6 220318
 	* Changes to graph modes due to matplotlib upgrade 2.2.2 causing display issues
 	* -a data mode and -n notify mode added.
+
+* Version 2.0-1 040418
+	* number of graph modes increased to 6 from 2 with menu selection , cli option -G removed
+	* csv option added -s which converts log file to csv for third party use.
+	* Added CPU,Ram and swap memory usage data 
+	* mail data now sent as attachment